Uses namespace to categorize LnxInit tests
