+++ Tests/Runtime/LnxInit/AbstractParameterTests.cs
@@ -7,9 +7,9 @@ using LnxArch.TestFixtures;
 using UnityEngine;
 using UnityEngine.TestTools;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_AbstractParameterTests
+    public class AbstractParameterTests
     {
 
         #region Nested
