+++ Tests/Runtime/LnxInit/CircularImmediateTests.cs.meta
