+++ Tests/Runtime/LnxInit/CircularSelfTests.cs
@@ -8,44 +8,44 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using System.Linq;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_DefaultParameterTests
+    public class CircularSelfTests
     {
         #region Nested
         public struct EntityContext
         {
-            public CaseDefaultParameter Main { get; private set; }
-            public Collider Collider { get; private set; }
+            public CaseCircularSelf Main { get; private set; }
+            public CallsRegister Register { get; private set; }
 
             public static EntityContext Create(LnxEntity entity)
             {
                 return new EntityContext {
-                    Main = entity.FetchFirst<CaseDefaultParameter>(),
-                    Collider = entity.FetchFirst<Collider>()
+                    Main = entity.FetchFirst<CaseCircularSelf>(),
+                    Register = entity.FetchFirst<CallsRegister>(),
                 };
             }
         }
         #endregion
         private readonly FixturesLoader _fixtures = FixturesLoader.RuntimeLnxInit;
         [Test]
-        public void WhenComponentIsFound_InjectsIt()
+        public void InjectsItSelf()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("DefaultParameter_Found");
+            var entity = _fixtures.InstantiateEntityPrefab("CircularSelf");
             EntityContext ctx = EntityContext.Create(entity);
             Assert.That(ctx.Main, Is.Not.Null);
-            Assert.That(ctx.Collider, Is.Not.Null);
 
-            Assert.That(ctx.Main.Collider, Is.EqualTo(ctx.Collider));
+            Assert.That(ctx.Main.Circular, Is.EqualTo(ctx.Main));
         }
+
         [Test]
-        public void WhenComponentIsNotFound_UsesDefaultValue()
+        public void OnlyInitializesOnce()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("DefaultParameter_NotFound");
+            var entity = _fixtures.InstantiateEntityPrefab("CircularSelf");
             EntityContext ctx = EntityContext.Create(entity);
             Assert.That(ctx.Main, Is.Not.Null);
 
-            Assert.That(ctx.Main.Collider, Is.Null);
+            Assert.That(ctx.Register.Calls.Count, Is.EqualTo(1));
         }
     }
 }
