+++ Tests/Runtime/LnxInit/CircularSelfTests.cs.meta
