+++ Tests/Runtime/LnxInit/LookupFromEntityTests.cs
@@ -1,16 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using LnxArch;
 using NUnit.Framework;
 using LnxArch.TestTools;
 using LnxArch.TestFixtures;
 using UnityEngine;
+using UnityEngine.TestTools;
+using System.Linq;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_ObjectsArrayTests
+    public class LookupFromEntityTests
     {
         #region Nested
         public struct EntityContext
         {
-            public CaseObjectsArray Main { get; private set; }
+            public CaseLookupFromEntity Main { get; private set; }
             public BoxCollider BoxCollider { get; private set; }
             public SphereCollider SphereCollider { get; private set; }
             public CapsuleCollider CapsuleCollider { get; private set; }
@@ -18,7 +23,7 @@ namespace LnxArch.Tests
             public static EntityContext Create(LnxEntity entity)
             {
                 return new EntityContext {
-                    Main = entity.FetchFirst<CaseObjectsArray>(),
+                    Main = entity.FetchFirst<CaseLookupFromEntity>(),
                     BoxCollider = entity.FetchFirst<BoxCollider>(),
                     SphereCollider = entity.FetchFirst<SphereCollider>(),
                     CapsuleCollider = entity.FetchFirst<CapsuleCollider>()
@@ -28,30 +33,25 @@ namespace LnxArch.Tests
         #endregion
         private readonly FixturesLoader _fixtures = FixturesLoader.RuntimeLnxInit;
         [Test]
-        public void WhenHasMatch_ReturnsAllMatchingComponents()
+        public void WhenFetchingOne_GetTheFirstComponentFoundInTheEntity()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("ObjectsArray");
+            var entity = _fixtures.InstantiateEntityPrefab("LookupFromEntity");
             EntityContext ctx = EntityContext.Create(entity);
-            Assert.That(ctx.Main, Is.Not.Null);
-            Assert.That(ctx.BoxCollider, Is.Not.Null);
-            Assert.That(ctx.SphereCollider, Is.Not.Null);
-            Assert.That(ctx.CapsuleCollider, Is.Not.Null);
 
-            Assert.That(ctx.Main.Colliders, Is.EqualTo(new Collider[] {
-                ctx.BoxCollider,
-                ctx.SphereCollider,
-                ctx.CapsuleCollider
-            }));
+            Assert.That(ctx.Main.Collider, Is.EqualTo(ctx.BoxCollider));
         }
 
         [Test]
-        public void WhenHasNoMatch_ReturnsEmptyArray()
+        public void WhenFetchingMany_GetAllTheComponentsFoundInTheEntity()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("ObjectsArray");
+            var entity = _fixtures.InstantiateEntityPrefab("LookupFromEntity");
             EntityContext ctx = EntityContext.Create(entity);
-            Assert.That(ctx.Main, Is.Not.Null);
 
-            Assert.That(ctx.Main.Rigidbodies, Is.Empty);
+            Assert.That(ctx.Main.Colliders, Is.EqualTo(new List<Collider> {
+                ctx.BoxCollider,
+                ctx.SphereCollider,
+                ctx.CapsuleCollider,
+            }));
         }
     }
 }
