+++ Tests/Runtime/LnxInit/LookupFromLocalTests.cs
@@ -1,17 +1,21 @@
+using System.Collections;
 using System.Collections.Generic;
+using LnxArch;
 using NUnit.Framework;
 using LnxArch.TestTools;
 using LnxArch.TestFixtures;
 using UnityEngine;
+using UnityEngine.TestTools;
+using System.Linq;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_ObjectsListTests
+    public class LookupFromLocalTests
     {
         #region Nested
         public struct EntityContext
         {
-            public CaseObjectsList Main { get; private set; }
+            public CaseLookupFromLocal Main { get; private set; }
             public BoxCollider BoxCollider { get; private set; }
             public SphereCollider SphereCollider { get; private set; }
             public CapsuleCollider CapsuleCollider { get; private set; }
@@ -19,7 +23,7 @@ namespace LnxArch.Tests
             public static EntityContext Create(LnxEntity entity)
             {
                 return new EntityContext {
-                    Main = entity.FetchFirst<CaseObjectsList>(),
+                    Main = entity.FetchFirst<CaseLookupFromLocal>(),
                     BoxCollider = entity.FetchFirst<BoxCollider>(),
                     SphereCollider = entity.FetchFirst<SphereCollider>(),
                     CapsuleCollider = entity.FetchFirst<CapsuleCollider>()
@@ -29,30 +33,27 @@ namespace LnxArch.Tests
         #endregion
         private readonly FixturesLoader _fixtures = FixturesLoader.RuntimeLnxInit;
         [Test]
-        public void WhenHasMatch_ReturnsAllMatchingComponents()
+        public void WhenFetchingOne_GetTheFirstComponentFoundInTheLocalObject()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("ObjectsList");
+            var entity = _fixtures.InstantiateEntityPrefab("LookupFromLocal");
             EntityContext ctx = EntityContext.Create(entity);
             Assert.That(ctx.Main, Is.Not.Null);
             Assert.That(ctx.BoxCollider, Is.Not.Null);
-            Assert.That(ctx.SphereCollider, Is.Not.Null);
-            Assert.That(ctx.CapsuleCollider, Is.Not.Null);
 
-            Assert.That(ctx.Main.Colliders, Is.EqualTo(new List<Collider>() {
-                ctx.BoxCollider,
-                ctx.SphereCollider,
-                ctx.CapsuleCollider
-            }));
+            Assert.That(ctx.Main.Collider, Is.EqualTo(ctx.BoxCollider));
         }
 
         [Test]
-        public void WhenHasNoMatch_ReturnsEmptyArray()
+        public void WhenFetchingMany_GetAllTheComponentsFoundInTheEntity()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("ObjectsList");
+            var entity = _fixtures.InstantiateEntityPrefab("LookupFromLocal");
             EntityContext ctx = EntityContext.Create(entity);
             Assert.That(ctx.Main, Is.Not.Null);
+            Assert.That(ctx.BoxCollider, Is.Not.Null);
 
-            Assert.That(ctx.Main.Rigidbodies, Is.Empty);
+            Assert.That(ctx.Main.Colliders, Is.EqualTo(new List<Collider> {
+                ctx.BoxCollider,
+            }));
         }
     }
 }
