+++ Tests/Runtime/LnxInit/LookupOrderTests.cs
@@ -8,9 +8,9 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using System.Linq;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_LookupOrderTests
+    public class LookupOrderTests
     {
         #region Nested
         public struct EntityContext
