+++ Tests/Runtime/LnxInit/MethodWithoutParameters.cs.meta
