+++ Tests/Runtime/LnxInit/MultipleCallsTests.cs
@@ -3,9 +3,9 @@ using LnxArch.TestTools;
 using LnxArch.TestFixtures;
 using UnityEngine;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_MultipleCallsTests
+    public class MultipleCallsTests
     {
         #region Nested
         public struct EntityContext
