+++ Tests/Runtime/LnxInit/ObjectsArrayTests.cs
@@ -1,21 +1,16 @@
-using System.Collections;
-using System.Collections.Generic;
-using LnxArch;
 using NUnit.Framework;
 using LnxArch.TestTools;
 using LnxArch.TestFixtures;
 using UnityEngine;
-using UnityEngine.TestTools;
-using System.Linq;
 
-namespace LnxArch.Tests
+namespace LnxArch.LnxInitTests
 {
-    public class LnxInit_LookupFromEntityTests
+    public class ObjectsArrayTests
     {
         #region Nested
         public struct EntityContext
         {
-            public CaseLookupFromEntity Main { get; private set; }
+            public CaseObjectsArray Main { get; private set; }
             public BoxCollider BoxCollider { get; private set; }
             public SphereCollider SphereCollider { get; private set; }
             public CapsuleCollider CapsuleCollider { get; private set; }
@@ -23,7 +18,7 @@ namespace LnxArch.Tests
             public static EntityContext Create(LnxEntity entity)
             {
                 return new EntityContext {
-                    Main = entity.FetchFirst<CaseLookupFromEntity>(),
+                    Main = entity.FetchFirst<CaseObjectsArray>(),
                     BoxCollider = entity.FetchFirst<BoxCollider>(),
                     SphereCollider = entity.FetchFirst<SphereCollider>(),
                     CapsuleCollider = entity.FetchFirst<CapsuleCollider>()
@@ -33,25 +28,30 @@ namespace LnxArch.Tests
         #endregion
         private readonly FixturesLoader _fixtures = FixturesLoader.RuntimeLnxInit;
         [Test]
-        public void WhenFetchingOne_GetTheFirstComponentFoundInTheEntity()
+        public void WhenHasMatch_ReturnsAllMatchingComponents()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("LookupFromEntity");
+            var entity = _fixtures.InstantiateEntityPrefab("ObjectsArray");
             EntityContext ctx = EntityContext.Create(entity);
+            Assert.That(ctx.Main, Is.Not.Null);
+            Assert.That(ctx.BoxCollider, Is.Not.Null);
+            Assert.That(ctx.SphereCollider, Is.Not.Null);
+            Assert.That(ctx.CapsuleCollider, Is.Not.Null);
 
-            Assert.That(ctx.Main.Collider, Is.EqualTo(ctx.BoxCollider));
+            Assert.That(ctx.Main.Colliders, Is.EqualTo(new Collider[] {
+                ctx.BoxCollider,
+                ctx.SphereCollider,
+                ctx.CapsuleCollider
+            }));
         }
 
         [Test]
-        public void WhenFetchingMany_GetAllTheComponentsFoundInTheEntity()
+        public void WhenHasNoMatch_ReturnsEmptyArray()
         {
-            var entity = _fixtures.InstantiateEntityPrefab("LookupFromEntity");
+            var entity = _fixtures.InstantiateEntityPrefab("ObjectsArray");
             EntityContext ctx = EntityContext.Create(entity);
+            Assert.That(ctx.Main, Is.Not.Null);
 
-            Assert.That(ctx.Main.Colliders, Is.EqualTo(new List<Collider> {
-                ctx.BoxCollider,
-                ctx.SphereCollider,
-                ctx.CapsuleCollider,
-            }));
+            Assert.That(ctx.Main.Rigidbodies, Is.Empty);
         }
     }
 }
